requirements.txt
sentence-transformers==2.2.2
transformers==4.30.2
torch==1.13.1
tokenizers==0.13.3
huggingface_hub==0.14.1
numpy<2

PyMuPDF==1.21.1
pdf2image==1.16.3
Pillow==9.5.0
langdetect==1.0.9
pytesseract==0.3.10